+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "hakka",
-  "version": "1.0.11",
+  "version": "1.0.12",
   "private": true,
   "description": "Hakka",
   "author": "Thanh Tam",
