+++ src/containers/SideBar/SideBarItem/index.tsx
@@ -6,32 +6,14 @@ import { Box, Flex, Text } from 'rebass';
 import styles from './styles';
 import { upperCaseFirstLetter } from '../../../common/functions';
 import { NOTIFICATION_DOT } from '..';
-import { notificationMissionAddresses } from '../../../constants/challenge';
 
 const SideBarItem = (props, { location, data }) => {
   const {
-    icon, text, path, subIcon,
+    icon, text, path, subIcon, isViewAllNotifiedMission
   } = props;
   const [selectedNavPath, setSelectedNavPath] = useState('');
   const isBrowser = typeof window !== 'undefined';
   const currentPath = isBrowser ? window.location.pathname.replace(/\//g, '').split('0',1) : '';
-  const [isViewAllNotifiedMission, setIsViewAllNotifiedMission] = useState<boolean>(true)
-
-  useEffect(() => {
-    const localStorageViewedPages = isBrowser ? window.localStorage.getItem('viewed-pages') : ''
-    if (localStorageViewedPages) {
-      let isAllAddressPass = true
-      const viewedPages: string[] = JSON.parse(localStorageViewedPages)
-      notificationMissionAddresses.forEach((address: string) => {
-        if (viewedPages.findIndex((viewedAddress: string) => viewedAddress === address) === -1) {
-          isAllAddressPass = false
-        }
-      })
-      setIsViewAllNotifiedMission(isAllAddressPass)
-    } else if (notificationMissionAddresses.length > 0) {
-      setIsViewAllNotifiedMission(false)
-    }
-  }, [])
 
   useEffect(() => {
     setSelectedNavPath(currentPath[0]);
